Move script main methods to cp3_llbb.SAMADhi.scripts or the module they logically belong with, and use entry_points/console_scripts
